refactor(Player) : 딜러와 비교하여 플레이어의 승부 결과를 반환하는 기능 이동
